Endret exceptionhåndteringen både i div routes, men også innover i service og domene-laget. Ingen grunn til å returnere null i de tilfeller hvor vi uansett sjekker på null og kaster exception rett etterpå, bedre å kaste ex med en gang
