+++ app/src/main/kotlin/no/nav/tiltakspenger/vedtak/routes/rivers/InnsendingUtdatertRoutes.kt
@@ -25,8 +25,7 @@ fun Route.innsendingUtdatertRoutes(
 ) {
     post(innsendingUtdatertRoute) {
         LOG.info { "Vi har mottatt InnsendingUtdatert fra river" }
-        val systembruker: Systembruker = innloggetSystembrukerProvider.hentInnloggetSystembruker(call)
-            ?: return@post call.respond(message = "JWTToken ikke funnet", status = HttpStatusCode.Unauthorized)
+        val systembruker: Systembruker = innloggetSystembrukerProvider.krevInnloggetSystembruker(call)
 
         LOG.info { "Vi ble kallt med systembruker : $systembruker" }
