+++ app/src/main/kotlin/no/nav/tiltakspenger/vedtak/routes/rivers/PersonopplysningerRoutes.kt
@@ -51,9 +51,7 @@ fun Route.personopplysningerRoutes(
 ) {
     post("$personopplysningerPath") {
         LOG.info { "Vi har mottatt personopplysninger fra river" }
-        val systembruker: Systembruker = innloggetSystembrukerProvider.hentInnloggetSystembruker(call)
-            ?: return@post call.respond(message = "JWTToken ikke funnet", status = HttpStatusCode.Unauthorized)
-
+        val systembruker: Systembruker = innloggetSystembrukerProvider.krevInnloggetSystembruker(call)
         LOG.info { "Vi ble kallt med systembruker : $systembruker" }
 
         val personopplysningerMottattDTO = call.receive<PersonopplysningerMottattDTO>()
