+++ app/src/main/kotlin/no/nav/tiltakspenger/vedtak/service/ports/BehandlingRepo.kt
@@ -8,7 +8,8 @@ import no.nav.tiltakspenger.felles.SakId
 interface BehandlingRepo {
     fun lagre(behandling: Førstegangsbehandling): Førstegangsbehandling
     fun lagre(behandling: Førstegangsbehandling, tx: TransactionalSession): Førstegangsbehandling
-    fun hent(behandlingId: BehandlingId): Førstegangsbehandling?
+    fun hentOrNull(behandlingId: BehandlingId): Førstegangsbehandling?
+    fun hent(behandlingId: BehandlingId): Førstegangsbehandling
     fun hentAlle(): List<Førstegangsbehandling>
     fun hentAlleForIdent(ident: String): List<Førstegangsbehandling>
     fun hentForSak(sakId: SakId): List<Førstegangsbehandling>
