+++ app/src/main/kotlin/no/nav/tiltakspenger/vedtak/service/søker/SøkerServiceImpl.kt
@@ -1,18 +1,20 @@
 package no.nav.tiltakspenger.vedtak.service.søker
 
 import mu.KotlinLogging
+import no.nav.tiltakspenger.exceptions.IkkeFunnetException
 import no.nav.tiltakspenger.felles.Saksbehandler
 import no.nav.tiltakspenger.felles.SøkerId
 import no.nav.tiltakspenger.vedtak.innsending.Søker
 import no.nav.tiltakspenger.vedtak.repository.søker.SøkerRepository
 
 private val LOG = KotlinLogging.logger {}
+val SECURELOG = KotlinLogging.logger("tjenestekall")
 
 class SøkerServiceImpl(
     private val søkerRepository: SøkerRepository,
 ) : SøkerService {
 
-    override fun hentSøkerId(ident: String, saksbehandler: Saksbehandler): SøkerIdDTO? {
+    override fun hentSøkerIdOrNull(ident: String, saksbehandler: Saksbehandler): SøkerIdDTO? {
         val søker: Søker = søkerRepository.findByIdent(ident) ?: return null
         søker.sjekkOmSaksbehandlerHarTilgang(saksbehandler)
         return SøkerIdDTO(
@@ -20,9 +22,21 @@ class SøkerServiceImpl(
         )
     }
 
-    override fun hentIdent(søkerId: SøkerId, saksbehandler: Saksbehandler): String? {
+    override fun hentIdentOrNull(søkerId: SøkerId, saksbehandler: Saksbehandler): String? {
         val søker: Søker = søkerRepository.hent(søkerId) ?: return null
         søker.sjekkOmSaksbehandlerHarTilgang(saksbehandler)
         return søker.ident
+    }
+
+    override fun hentSøkerId(ident: String, saksbehandler: Saksbehandler): SøkerIdDTO {
+        return hentSøkerIdOrNull(ident, saksbehandler)
+            ?: throw IkkeFunnetException("SøkerId ikke funnet")
+                .also { SECURELOG.warn { "SøkerId ikke funnet for ident $ident og saksbehandler $saksbehandler" } }
+    }
+
+    override fun hentIdent(søkerId: SøkerId, saksbehandler: Saksbehandler): String {
+        return hentIdentOrNull(søkerId, saksbehandler)
+            ?: throw IkkeFunnetException("Ident ikke funnet")
+                .also { SECURELOG.warn { "Ident ikke funnet for søkerId $søkerId og saksbehandler $saksbehandler" } }
     }
 }
