+++ app/src/main/kotlin/no/nav/tiltakspenger/vedtak/tilgang/InnloggetSaksbehandlerProvider.kt
@@ -4,7 +4,7 @@ import io.ktor.server.application.ApplicationCall
 import io.ktor.server.auth.jwt.JWTPrincipal
 import io.ktor.server.auth.principal
 import no.nav.tiltakspenger.felles.Saksbehandler
-import no.nav.tiltakspenger.vedtak.routes.exceptionhandling.ManglendeJWTTokenException
+import no.nav.tiltakspenger.vedtak.exceptions.ManglendeJWTTokenException
 
 interface InnloggetSaksbehandlerProvider {
     fun hentSaksbehandler(principal: JWTPrincipal): Saksbehandler
