+++ domene/src/main/kotlin/no/nav/tiltakspenger/exceptions/IkkeFunnetException.kt
@@ -1,0 +1,3 @@
+package no.nav.tiltakspenger.exceptions
+
+class IkkeFunnetException(override val message: String) : RuntimeException(message)
