+++ domene/src/test/kotlin/no/nav/tiltakspenger/felles/UlidTest.kt
@@ -12,14 +12,14 @@ internal class UlidTest {
     @Test
     fun `test roundtrip`() {
         val ulid = BehandlingId.random()
-        val ulid2 = BehandlingId.fromDb(ulid.toString())
+        val ulid2 = BehandlingId.fromString(ulid.toString())
         assertEquals(ulid, ulid2)
     }
 
     @Test
     fun `test prefixPart and ulidPart`() {
         val ulid = BehandlingId.random()
-        val ulid2 = BehandlingId.fromDb("${ulid.prefixPart()}_${ulid.ulidPart()}")
+        val ulid2 = BehandlingId.fromString("${ulid.prefixPart()}_${ulid.ulidPart()}")
         assertEquals(ulid, ulid2)
     }
 
@@ -27,27 +27,27 @@ internal class UlidTest {
     fun `test fromDb negativ test`() {
         val utenGyldigSkilletegn = "HH"
         shouldThrowWithMessage<IllegalArgumentException>("Ikke gyldig Id, skal bestå av to deler skilt med _") {
-            BehandlingId.fromDb(utenGyldigSkilletegn)
+            BehandlingId.fromString(utenGyldigSkilletegn)
         }
 
         val utenPrefiks = "_HH"
         shouldThrowWithMessage<IllegalArgumentException>("Ikke gyldig Id, prefiks er tom") {
-            BehandlingId.fromDb(utenPrefiks)
+            BehandlingId.fromString(utenPrefiks)
         }
 
         val utenUlid = "HH_"
         shouldThrowWithMessage<IllegalArgumentException>("Ikke gyldig Id, ulid er ugyldig") {
-            BehandlingId.fromDb(utenUlid)
+            BehandlingId.fromString(utenUlid)
         }
 
         val ugyldeTegniUlid = "HH_1234567890123456789U123456"
         shouldThrowWithMessage<IllegalArgumentException>("Ikke gyldig Id, ulid er ugyldig") {
-            BehandlingId.fromDb(ugyldeTegniUlid)
+            BehandlingId.fromString(ugyldeTegniUlid)
         }
 
         val ugyldigUlid = "HH_UU_JJ"
         shouldThrowWithMessage<IllegalArgumentException>("Ikke gyldig Id, skal bestå av prefiks + ulid") {
-            BehandlingId.fromDb(ugyldigUlid)
+            BehandlingId.fromString(ugyldigUlid)
         }
     }
